Use NodePrimaryIp to set host-addresses in DPU mode
When running a 2-cluster design in DPU mode, currently host-addresses
is set to the IP address of the Infrastructure node (i.e. Bluefield
node) running of behalf of the tenant cluster. This will cause all
NodePort service to fail in DPU mode.

Instead, we should calculate the host-addresses using the NodePrimaryIP
which will set the host-address to the Node IP of the tenant node.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>